+++ packages/editor/editor-core/src/types/editor-props.ts
@@ -9,7 +9,6 @@ import {
   ExtensionHandlers,
 } from '@atlaskit/editor-common';
 import { ActivityProvider } from '@atlaskit/activity';
-import { DelegateAnalyticsEvent } from '@atlaskit/analytics';
 import { MentionProvider } from '@atlaskit/mention';
 import { EmojiProvider } from '@atlaskit/emoji';
 import { TaskDecisionProvider } from '@atlaskit/task-decision';
@@ -74,9 +73,6 @@ export interface EditorProps {
   // Legacy analytics support handler, which will be removed soon. **Do not use**.
   analyticsHandler?: AnalyticsHandler;
 
-  // For `@atlaskit/analytics` support
-  delegateAnalyticsEvent?: DelegateAnalyticsEvent;
-
   contentComponents?: ReactComponents;
   primaryToolbarComponents?: ReactComponents;
   secondaryToolbarComponents?: ReactComponents;
@@ -113,7 +109,7 @@ export interface EditorProps {
 
   // Set this to allow unsupported content in the editor.
   // Anything it doesn’t understand it will wrap in an unsupported block or inline node.
-  // It will render a grey non editable box.
+  // It will render a gray non editable box.
   allowUnsupportedContent?: boolean;
 
   // Enable panel blocks, the thing that displays a coloured box with icons aka info, warning macros.