+++ packages/media/media-core/examples/1-get-file.tsx
@@ -62,14 +62,14 @@ class Example extends Component<ComponentProps, ComponentState> {
   };
 
   getFile = (id: string, collectionName?: string) => {
-    const stream = mediaContext.getFile(id, { collectionName });
+    const stream = mediaContext.file.getFileState(id, { collectionName });
 
     this.addStream(stream);
   };
 
   uploadFile = async (event: SyntheticEvent<HTMLInputElement>) => {
     const file = event.currentTarget.files![0];
-    const stream = mediaContext.uploadFile({
+    const stream = mediaContext.file.upload({
       content: file,
       name: file.name,
       collection: defaultCollectionName,