+++ packages/media/media-test-helpers/src/fakeContextProvider.ts
@@ -41,7 +41,10 @@ export const fakeContext = (
   const collection = {
     loadNextPage: jest.fn(),
   } as any;
-  const getImage = {} as any;
+  const getImage = jest.fn() as any;
+  const file = {
+    getFileState: getFile,
+  } as any;
   const defaultContext: Context = {
     getImage,
     getFile,
@@ -58,12 +61,29 @@ export const fakeContext = (
     uploadFile,
     config,
     collection,
+    file,
   };
 
   const wrappedStubbedContext: any = {};
   Object.keys(stubbedContext).forEach(methodName => {
     wrappedStubbedContext[methodName] = returns(stubbedContext[methodName]);
   });
+
+  if (stubbedContext.file) {
+    Object.keys(stubbedContext.file).forEach(methodName => {
+      wrappedStubbedContext.file[methodName] = returns(
+        stubbedContext.file[methodName],
+      );
+    });
+  }
+
+  if (stubbedContext.context) {
+    Object.keys(stubbedContext.context).forEach(methodName => {
+      wrappedStubbedContext.context[methodName] = returns(
+        stubbedContext.context[methodName],
+      );
+    });
+  }
 
   return {
     ...defaultContext,