+++ packages/search/global-search/src/__tests__/unit/ResultList.test.tsx
@@ -10,6 +10,7 @@ import PageIcon from '@atlaskit/icon-object/glyph/page/24';
 import BoardIcon from '@atlaskit/icon/glyph/board';
 import ResultList, { Props } from '../../components/ResultList';
 import {
+  Result,
   JiraResult,
   PersonResult,
   AnalyticsType,
@@ -138,6 +139,36 @@ it('should pass the correct properties to ContainerResult for Confluence spaces'
     name: 'name',
     analyticsData: expect.objectContaining(DUMMY_ANALYTICS_DATA),
   });
+});
+
+it('should avoid duplicate result keys', () => {
+  const results: Result[] = [
+    makeConfluenceContainerResult({
+      resultId: 'resultId',
+      analyticsType: AnalyticsType.ResultConfluence,
+    }),
+    makeConfluenceObjectResult({
+      resultId: 'resultId',
+      analyticsType: AnalyticsType.ResultConfluence,
+    }),
+    makePersonResult({
+      resultId: 'resultId',
+      analyticsType: AnalyticsType.ResultPerson,
+    }),
+  ];
+
+  const wrapper = render({
+    results,
+    analyticsData: DUMMY_ANALYTICS_DATA,
+  });
+
+  const containerKey = wrapper.find(ContainerResultComponent).key();
+  const objectKey = wrapper.find(ObjectResultComponent).key();
+  const personKey = wrapper.find(PersonResultComponent).key();
+
+  expect(containerKey).toBe('confluence-space-resultId-1');
+  expect(objectKey).toBe('confluence-page-resultId-2');
+  expect(personKey).toBe('result-resultId-3');
 });
 
 describe('Jira Avatar default Icons', () => {